+++ configs/config_benchmark.yaml
@@ -5,7 +5,7 @@ general:
   app_name: "Pulse"
   app_version: "1.0.0"
   app_mode: "debug" # "debug", "count_tokens", "benchmark",
-  debug_data_length: 100 # Number of samples to use for debugging
+  debug_data_length: 10 # Number of samples to use for debugging
   logging_level: "DEBUG" # "DEBUG", "INFO", "WARNING", "ERROR"
   use_scratch: true # Whether to use scratch space for temporary files
   save_metadata: true # Whether to save the metadata file. Inludes prompts for LLMs
@@ -36,14 +36,14 @@ datasets:
 preprocessing_baseline:
   replace_outliers: true # Whether to replace outliers with NaNs
   flag_na: true # Whether to flag NA values with binary indicators
-  standardize: true # Whether to standardize features (yes: convDL, not required: convML, LLM)
+  standardize: false # Whether to standardize features (yes: convDL, not required: convML, LLM)
   static_imputation: true # Whether to perform static feature imputation
   dynamic_imputation: true # Whether to perform dynamic feature imputation
   split_ratios: # Train/Val/Test split ratios
     train: 0.8
     val: 0.1
     test: 0.1
-    test_limited: 100 # Number of stay_ids for limited test set, null = test set is not limited
+    test_limited: 10 # Number of stay_ids for limited test set, null = test set is not limited
     print_stats: true # Whether to print set statistics
 
 # _______ Preprocessing Advanced _______
