+++ src/models/gemma3_model.py
@@ -20,7 +20,7 @@ from transformers import (
 
 import wandb
 from src.eval.metrics import MetricsTracker
-from src.models.pulsetemplate_model import PulseTemplateModel
+from src.models.pulse_model import PulseTemplateModel
 from src.util.model_util import extract_dict, prompt_template_hf
 from src.util.config_util import set_seeds
